Better organisation of pick dataset tool
